refactor. 필터 API 요청 DTO 패키지 변경
