+++ BE/src/main/java/team07/airbnb/service/accommodation/AccommodationService.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import team07.airbnb.common.util.DateHelper;
 import team07.airbnb.common.util.GeometryHelper;
-import team07.airbnb.controller.AccommodationFilterDTO;
+import team07.airbnb.data.accommodation.dto.request.AccommodationFilterDTO;
 import team07.airbnb.data.accommodation.enums.AccommodationType;
 import team07.airbnb.data.user.enums.Role;
 import team07.airbnb.entity.AccommodationEntity;
