+++ BE/src/main/java/team07/airbnb/service/product/ProductService.java
@@ -3,13 +3,12 @@ package team07.airbnb.service.product;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import team07.airbnb.controller.AccommodationFilterDTO;
+import team07.airbnb.data.accommodation.dto.request.AccommodationFilterDTO;
 import team07.airbnb.data.accommodation.dto.response.AccommodationListResponse;
 import team07.airbnb.data.product.dto.response.ProductListResponse;
 import team07.airbnb.entity.AccommodationEntity;
 import team07.airbnb.entity.BookingEntity;
 import team07.airbnb.entity.ProductEntity;
-import team07.airbnb.entity.UserEntity;
 import team07.airbnb.exception.IllegalRequestException;
 import team07.airbnb.exception.auth.UnAuthorizedException;
 import team07.airbnb.exception.not_found.ProductNotFoundException;
